% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{analyze - an over-the-top word frequency counter}
\usage{
analyze(textIn = "file", keepHyphens = TRUE, keepContractions = TRUE)
}
\arguments{
\item{keepContractions}{}
}
\value{
your.variable <- c("word" = #count, "etc" = #, ...)
}
\description{
analyze - an over-the-top word frequency counter
}
\examples{
wordCounts <- analyze()
wordCounts <- analyze("don't look at this text", keepContractions=FALSE)
wordCounts <- analyze(textIn="don't enter file-paths yet", keepHyphens=FALSE)

}
