% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter}
\alias{counter}
\title{Counter parses the text input and creates a simple named vector to hold
int counts of words. Accepts string input but defaults to choose a file.}
\usage{
counter(textIn = chooseFile, keepHyphens = TRUE, keepContractions = TRUE)
}
\arguments{
\item{textIn}{defaults to chooseFile or input a string}

\item{keepHyphens}{(TRUE: "single-file" += 1 |
                     FALSE: "single" += 1, "file" += 1)
note:  counter does not keep end-of-line hyphens; it recombines the word
       Does not recognize double-hyphens; will falsely recombine single-hyphens}

\item{keepContractions}{(TRUE:"can't" += 1 |FALSE: "can" += 1, "t" += 1)}
}
\value{
named vector
}
\description{
Counter parses the text input and creates a simple named vector to hold
int counts of words. Accepts string input but defaults to choose a file.
}
\examples{
counter()
counter("Here-here is some text, here!", keepHyphens = FALSE)
counter("What isn't this doing to my text?! Don't do that!", keepContractions = FALSE)
}
