% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factotum.R
\name{factotum}
\alias{factotum}
\title{factotum - a person who does a variety of tasks}
\usage{
factotum(namedVector = analyze(), top = 10, bottom = 10, hapaxLegomena = TRUE)
}
\arguments{
\item{namedVector}{defaults to use analyze}

\item{top}{t most common used words}

\item{bottom}{b least common words (alphabetized)}

\item{hapaxLegomena}{shows all the unique words (obviates bottom argument)}
}
\value{
nothing, prints a report
}
\description{
factotum - a person who does a variety of tasks
}
\examples{
factotum(hapaxLegomena = FALSE)
factotum(c("find" = 10, "the" = 3, "top" = 2), top = 2)
}
