% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salient.R
\name{salient}
\alias{salient}
\title{salient - basically the opposite of quotidian}
\usage{
salient(namedVector = analyze(), show = 10, common = TRUE)
}
\arguments{
\item{namedVector}{defaults to analyze()}

\item{common}{display frequent words? (T/F)}

\item{bottom}{number of lest-used words to show}
}
\value{
nothing, prints a list
}
\description{
salient - basically the opposite of quotidian
}
\examples{
salient()
salient(bottom = 15)
salient(vector)
}
