% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TexTesT.R
\docType{data}
\name{TexTesT}
\alias{TexTesT}
\title{This is pretty much just a text file with a back story for the curious.
It exists entirely to test the functions out and see how they behave.}
\format{
text
}
\usage{
TexTesT
}
\description{
This is pretty much just a text file with a back story for the curious.
It exists entirely to test the functions out and see how they behave.
}
\keyword{datasets}
